+++ packages/hix/lib/Hix/Env.hs
@@ -5,7 +5,12 @@ import Path (Abs, Dir, Path)
 
 import Hix.Component (targetComponent)
 import qualified Hix.Data.ComponentConfig
-import Hix.Data.ComponentConfig (EnvRunner (EnvRunner), PackagesConfig, TargetOrDefault (DefaultTarget, ExplicitTarget))
+import Hix.Data.ComponentConfig (
+  EnvRunner (EnvRunner),
+  PackageName,
+  PackagesConfig,
+  TargetOrDefault (DefaultTarget, ExplicitTarget),
+  )
 import Hix.Data.Error (pathText)
 import qualified Hix.Data.GhciConfig
 import Hix.Json (jsonConfig)
@@ -17,11 +22,12 @@ import Hix.Options (EnvRunnerOptions, TargetSpec)
 -- Nothing when the config requests it
 componentRunner ::
   Maybe (Path Abs Dir) ->
+  Maybe PackageName ->
   PackagesConfig ->
   TargetSpec ->
   M (Maybe EnvRunner)
-componentRunner cliRoot config spec =
-  targetComponent cliRoot config spec <&> \case
+componentRunner cliRoot defaultPkg config spec =
+  targetComponent cliRoot defaultPkg config spec <&> \case
     ExplicitTarget t -> t.component.runner
     DefaultTarget t -> t.component.runner
     _ -> Nothing
@@ -29,7 +35,8 @@ componentRunner cliRoot config spec =
 envRunner :: EnvRunnerOptions -> M EnvRunner
 envRunner opts = do
   config <- either pure jsonConfig opts.config
-  fromMaybe config.defaultEnv . join <$> traverse (componentRunner opts.root config.packages) opts.component
+  let runner = componentRunner opts.root config.mainPackage config.packages
+  fromMaybe config.defaultEnv . join <$> traverse runner opts.component
 
 printEnvRunner :: EnvRunnerOptions -> M ()
 printEnvRunner opts = do