+++ modules/basic.nix
@@ -309,7 +309,7 @@ in {
 
       relativePackages = util.relativePackages config.base config.internal.packagePaths;
 
-      hixVersion = "0.6.6";
+      hixVersion = "0.6.7";
 
     };
   };